change rational format from using int to long

+++ src/soplex_interface.h
@@ -19,11 +19,33 @@ void SoPlex_setIntParam(void *soplex, int paramcode, int paramvalue);
 
 void SoPlex_addColReal(void *soplex, double* colentries, int colsize, int nnonzeros, double objval, double lb, double ub);
 
-void SoPlex_addColRational(void *soplex, int* colnums, int* coldenoms, int colsize, int nnonzeros, int objvalnum, int objvaldenom, int lbnum, int lbdenom, int ubnum, int ubdenom);
+void SoPlex_addColRational(
+        void *soplex,
+        long* colnums,
+        long* coldenoms,
+        int colsize,
+        int nnonzeros,
+        long objvalnum,
+        long objvaldenom,
+        long lbnum,
+        long lbdenom,
+        long ubnum,
+        long ubdenom
+);
 
 void SoPlex_addRowReal(void *soplex, double* rowentries, int rowsize, int nnonzeros, double lb, double ub);
 
-void SoPlex_addRowRational(void *soplex, int* rownums, int* rowdenoms, int rowsize, int nnonzeros, int lbnum, int lbdenom, int ubnum, int ubdenom);
+void SoPlex_addRowRational(
+        void *soplex,
+        long* rownums,
+        long* rowdenoms,
+        int rowsize,
+        int nnonzeros,
+        long lbnum,
+        long lbdenom,
+        long ubnum,
+        long ubdenom
+);
 
 void SoPlex_getPrimalReal(void *soplex, double* primal, int dim);
 
@@ -35,15 +57,15 @@ int SoPlex_optimize(void *soplex);
 
 void SoPlex_changeObjReal(void *soplex, double* obj, int dim);
 
-void SoPlex_changeObjRational(void *soplex, int* objnums, int* objdenoms, int dim);
+void SoPlex_changeObjRational(void *soplex, long* objnums, long* objdenoms, int dim);
 
 void SoPlex_changeLhsReal(void *soplex, double* lhs, int dim);
 
-void SoPlex_changeLhsRational(void *soplex, int* lhsnums, int* lhsdenoms, int dim);
+void SoPlex_changeLhsRational(void *soplex, long* lhsnums, long* lhsdenoms, int dim);
 
 void SoPlex_changeRhsReal(void *soplex, double* rhs, int dim);
 
-void SoPlex_changeRhsRational(void *soplex, int* rhsnums, int* rhsdenoms, int dim);
+void SoPlex_changeRhsRational(void *soplex, long* rhsnums, long* rhsdenoms, int dim);
 
 void SoPlex_writeFileReal(void *soplex, char* filename);
 
@@ -55,7 +77,7 @@ void SoPlex_changeBoundsReal(void *soplex, double* lb, double* ub, int dim);
 
 void SoPlex_changeVarBoundsReal(void *soplex, int colidx, double lb, double ub);
 
-void SoPlex_changeVarBoundsRational(void *soplex, int colidx, int lbnum, int lbdenom, int ubnum, int ubdenom);
+void SoPlex_changeVarBoundsRational(void *soplex, int colidx, long lbnum, long lbdenom, long ubnum, long ubdenom);
 
 void SoPlex_changeVarUpperReal(void *soplex, int colidx, double ub);
